Update configure.sh
Better help message. Fault checking


Former-commit-id: 0a18f9a45f4ef85413086e170d53c527e25735bc

+++ demos/demo1/configure.sh
@@ -1,6 +1,9 @@
 #!/usr/bin/env bash
 
-source conf/demo.conf
+CONF_DIR=conf
+ROOT_DIR=`git rev-parse --show-toplevel`
+DEMO_DIR="$ROOT_DIR/demos/demo1"
+source $CONF_DIR/demo.conf
 REGISTRY=$REGISTRY_IP:$REGISTRY_PORT
 
 awk_script=" {
@@ -17,14 +20,55 @@ generate() {
 	mkdir -p yamls/security
 	mkdir -p yamls/pods
 	for file in $KUBE_DEPLOYMENTS; do
-		awk "$awk_script" conf/$file.yaml.conf > yamls/$file.yaml
+		awk "$awk_script" $CONF_DIR/$file.yaml.conf > yamls/$file.yaml
 	done;
 }
 
+check() {
+	if [ ! -d $CONF_DIR ]; then
+		echo "Error: Directory $CONF_DIR does not exist!"
+		echo "  Please make sure you are in $DEMO_DIR directory and"
+		echo "  your repo is up to date."
+		exit 2
+	fi
+	for file in $KUBE_DEPLOYMENTS; do
+		if [ ! -f $CONF_DIR/$file.yaml.conf ]; then
+			echo "Error: file $CONF_DIR/$file.yaml.conf not found!"
+			exit 2
+		fi
+	done;
+}
+
+# this funcrion cleans all yaml files generated from this script
+# that where secified in demo.conf. If you want to clean files 
+# generated with --file option you should remove them manualy,
+# or add them in KUBE_DEPLOYMENTS in demo.conf
 clean() {
 	for file in $KUBE_DEPLOYMENTS; do
+		if [ ! -f $file ]; then
+			continue
+		fi
 		rm yamls/$file.yaml 2> /dev/null
 	done;
+
+	if [ -d yamls/security ]; then
+		if [ ! "$(ls yamls/security 2> /dev/null)" ] ; then
+			echo "cleaning security"
+			rm -rf yamls/security 2> /dev/null
+		fi
+	fi
+	if [ -d yamls/pods ]; then
+		if [ ! "$(ls yamls/pods 2> /dev/null)" ] ; then
+			echo "cleaning pods"
+			rm -rf yamls/pods 2> /dev/null
+		fi
+	fi
+	if [ -d yamls ]; then
+		if [ ! "$(ls yamls 2> /dev/null)" ] ; then
+			echo "removing yamls"
+			rm -rf yamls 2> /dev/null
+		fi
+	fi
 }
 
 usage() {
@@ -32,15 +76,19 @@ usage() {
 	echo "       $1 [OPTION]"
 	echo "  Generates yaml deployment files for ddos demo"
 	echo ""
-	echo ""
 	echo "Options:"
-	echo "  --clean             remove generated yaml files"
-	echo "  --file FILENAME     generate for specific file"
+	echo "  --clean             remove generated yaml files. Files not defined in "
+	echo "                      KUBE_DEPLOYMENTS will NOT be removed"
+	echo "  --file FILENAME     generate for specified file FILENAME sould be in a form"
+	echo "                      <pods|security|other_custom_floder>/<filename_without_extention>"
+	echo "                      e.g.: ./configure --file pods/alice"
+	echo "                      If a custom file is used, it should have .yaml.conf extention"
 	echo "  --help              print this message"
 
 	exit $2
 }
 
+check
 if [ $# -eq 1 ]; then
 	if [ $1 = "--clean" ]; then
 		clean
